add step at chat room page
